src/requirements.txt
aiogram==3.25.0
psycopg2-binary==2.9.11
